Mega refactor of `handle_client`
Add a bunch of support functions to better present the flow of the request/response handling.

A list of changes:

- Rename `client` to `server`
- Move `start` and `servers` to a new file `start.rs`
- Add `get_request`
- Rework `get_route` to return `Result`
- Rework `get_method` to return `Result`
- Add `handle_(un)safe_request` which handles the request depending on if it is unsafe or not
- `handle_client` now returns `io::Result`
